fix seed

+++ db/seeds.rb
@@ -6,7 +6,9 @@
 # #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 # #   Character.create(name: 'Luke', movie: movies.first)
 
-User.find_by(email: "jul33anc@gmail.com").portfolio.destroy
+if User.find_by(email: "jul33anc@gmail.com").portfolio
+  User.find_by(email: "jul33anc@gmail.com").portfolio.destroy
+end
 User.find_by(email: "jul33anc@gmail.com").destroy
 
 Currency.update_data(200)
